Move event dispatcher internals
