+++ src/main/kotlin/io/github/freya022/botcommands/api/core/EventDispatcher.kt
@@ -1,341 +1,25 @@
 package io.github.freya022.botcommands.api.core
 
-import dev.minn.jda.ktx.events.CoroutineEventManager
 import io.github.freya022.botcommands.api.core.annotations.BEventListener
-import io.github.freya022.botcommands.api.core.annotations.BEventListener.RunMode
-import io.github.freya022.botcommands.api.core.config.BConfig
-import io.github.freya022.botcommands.api.core.config.BCoroutineScopesConfig
-import io.github.freya022.botcommands.api.core.events.BGenericEvent
-import io.github.freya022.botcommands.api.core.events.InitializationEvent
-import io.github.freya022.botcommands.api.core.service.ServiceContainer
-import io.github.freya022.botcommands.api.core.service.annotations.BService
-import io.github.freya022.botcommands.api.core.service.annotations.ServiceType
-import io.github.freya022.botcommands.api.core.utils.findAnnotationRecursive
-import io.github.freya022.botcommands.api.core.utils.isSubclassOf
-import io.github.freya022.botcommands.api.core.utils.simpleNestedName
-import io.github.freya022.botcommands.internal.core.*
-import io.github.freya022.botcommands.internal.core.exceptions.InternalException
-import io.github.freya022.botcommands.internal.core.service.FunctionAnnotationsMap
-import io.github.freya022.botcommands.internal.core.service.getParameters
-import io.github.freya022.botcommands.internal.utils.*
-import io.github.freya022.botcommands.internal.utils.ReflectionUtils.declaringClass
-import io.github.freya022.botcommands.internal.utils.ReflectionUtils.nonInstanceParameters
-import io.github.oshai.kotlinlogging.KotlinLogging
-import kotlinx.coroutines.*
-import net.dv8tion.jda.api.events.Event
-import net.dv8tion.jda.api.events.GenericEvent
-import net.dv8tion.jda.api.hooks.IEventManager
-import net.dv8tion.jda.api.requests.GatewayIntent
-import org.springframework.context.annotation.Bean
-import org.springframework.stereotype.Component
-import java.lang.reflect.InvocationTargetException
-import java.util.concurrent.ConcurrentHashMap
-import java.util.concurrent.locks.ReentrantLock
-import kotlin.concurrent.withLock
-import kotlin.reflect.KClass
-import kotlin.reflect.full.callSuspend
-import kotlin.reflect.full.functions
-import kotlin.reflect.jvm.jvmErasure
-import kotlin.time.Duration
-import kotlin.time.toDuration
-import kotlin.time.toDurationUnit
-
-internal class SortedList<T>(private val comparator: Comparator<T>) {
-    // Only protect modification operations, traversal is fine
-    private val lock = ReentrantLock()
-    private var list: List<T> = emptyList()
-
-    fun isEmpty() = list.isEmpty()
-    fun isNotEmpty() = !isEmpty()
-
-    fun add(t: T): Unit = lock.withLock {
-        val newList = list + t
-        this.list = newList.sortedWith(comparator)
-    }
-
-    fun remove(t: T): Boolean = lock.withLock {
-        val newList = list.toMutableList()
-        newList.remove(t).also {
-            this.list = newList
-        }
-    }
-
-    inline fun <R> map(block: (T) -> R): List<R> = list.map(block)
-
-    fun removeAll(removedList: SortedList<T>): Boolean = lock.withLock {
-        val newList = list.toMutableList()
-        return newList.removeAll(removedList.list).also {
-            this.list = newList
-        }
-    }
-
-    inline fun forEach(block: (T) -> Unit) = list.forEach(block)
-}
-
-private typealias EventMap = MutableMap<KClass<*>, SortedList<EventHandlerFunction>>
-
-private val logger = KotlinLogging.logger { }
-
-@BService // Cannot be a BConfiguration as the function below needs an instance
-@Component
-internal open class EventHooksProvider(
-    eventManagerSupplier: ICoroutineEventManagerSupplier,
-) {
-    private val originalCoroutineEventManager = eventManagerSupplier.get()
-
-    @Bean
-    @BService
-    @ServiceType(IEventManager::class)
-    internal open fun coroutineEventManager(
-        eventDispatcher: EventDispatcher,
-    ): CoroutineEventManager {
-        return DispatcherAwareCoroutineEventManager(originalCoroutineEventManager, eventDispatcher)
-    }
-
-    @Bean
-    @BService
-    fun eventDispatcher(
-        config: BConfig,
-        coroutineScopesConfig: BCoroutineScopesConfig,
-        serviceContainer: ServiceContainer,
-        eventTreeService: EventTreeService,
-        jdaService: JDAService,
-        functionAnnotationsMap: FunctionAnnotationsMap,
-    ): EventDispatcher {
-        return EventDispatcher(config, coroutineScopesConfig, serviceContainer, originalCoroutineEventManager, eventTreeService, jdaService, functionAnnotationsMap)
-    }
-}
-
-internal class DispatcherAwareCoroutineEventManager internal constructor(
-    originalCoroutineEventManager: CoroutineEventManager,
-    private val eventDispatcher: EventDispatcher,
-) : CoroutineEventManager(originalCoroutineEventManager, originalCoroutineEventManager.timeout) {
-
-    override fun handle(event: GenericEvent) {
-        eventDispatcher.onEvent(event)
-        super.handle(event) // Still let users use their own registered event listeners
-    }
-}
+import io.github.freya022.botcommands.api.core.service.annotations.InterfacedService
+import kotlinx.coroutines.Deferred
+import kotlinx.coroutines.runBlocking
 
 /**
  * Dispatches JDA and BC events to [@BEventListener][BEventListener] methods.
  */
-class EventDispatcher internal constructor(
-    private val config: BConfig,
-    private val coroutineScopesConfig: BCoroutineScopesConfig,
-    private val serviceContainer: ServiceContainer,
-    originalCoroutineEventManager: CoroutineEventManager,
-    private val eventTreeService: EventTreeService,
-    private val jdaService: JDAService,
-    functionAnnotationsMap: FunctionAnnotationsMap,
-) {
+@InterfacedService(acceptMultiple = false)
+abstract class EventDispatcher internal constructor() {
 
-    private val eventCoroutineScope: CoroutineScope = originalCoroutineEventManager
-    private val eventTimeout: Duration = originalCoroutineEventManager.timeout
+    abstract fun addEventListener(listener: Any)
 
-    private val map: EventMap = ConcurrentHashMap()
-    private val listeners: MutableMap<Class<*>, EventMap> = ConcurrentHashMap()
-
-    init {
-        functionAnnotationsMap
-            .get<BEventListener>()
-            .addAsEventListeners()
-    }
-
-    internal fun onEvent(event: GenericEvent) {
-        // No need to check for `event` type as if it's in the map, then it's recognized
-        val handlers = map[event::class] ?: return
-
-        // Run blocking handlers first
-        if (handlers.isNotEmpty()) {
-            runBlocking {
-                handlers.forEach { eventHandler ->
-                    if (eventHandler.runMode == RunMode.BLOCKING) {
-                        runEventHandler(eventHandler, event)
-                    }
-                }
-            }
-        }
-
-        // Stick to what JDA-KTX does, 1 coroutine per event for all listeners
-        eventCoroutineScope.launch {
-            handlers.forEach { eventHandler ->
-                if (eventHandler.runMode == RunMode.ASYNC) {
-                    coroutineScopesConfig.eventDispatcherScope.launch {
-                        runEventHandler(eventHandler, event)
-                    }
-                } else if (eventHandler.runMode == RunMode.INHERIT) {
-                    runEventHandler(eventHandler, event)
-                }
-            }
-        }
-    }
-
-    fun addEventListener(listener: Any) {
-        listener::class
-            .functions
-            .withFilter(FunctionFilter.annotation<BEventListener>())
-            .toClassPathFunctions(listener)
-            .addAsEventListeners()
-    }
-
-    fun removeEventListener(listener: Any) {
-        listeners.remove(listener::class.java)?.let { instanceMap ->
-            instanceMap.forEach { (kClass, functions) ->
-                val functionMap = map[kClass]
-                    ?: throwInternal("Listener was registered without having its functions added to the listener map")
-                if (!functionMap.removeAll(functions)) {
-                    logger.error(InternalException("Unable to remove listener functions from registered functions")) { "An exception occurred while removing event listener $listener" }
-                }
-            }
-        }
-    }
+    abstract fun removeEventListener(listener: Any)
 
     @JvmSynthetic
-    suspend fun dispatchEvent(event: Any) {
-        // No need to check for `event` type as if it's in the map, then it's recognized
-        val handlers = map[event::class] ?: return
-
-        // Run blocking handlers first
-        if (handlers.isNotEmpty()) {
-            runBlocking {
-                handlers.forEach { eventHandler ->
-                    if (eventHandler.runMode == RunMode.BLOCKING) {
-                        runEventHandler(eventHandler, event)
-                    }
-                }
-            }
-        }
-
-        handlers.forEach { eventHandler ->
-            if (eventHandler.runMode == RunMode.ASYNC) {
-                coroutineScopesConfig.eventDispatcherScope.launch {
-                    runEventHandler(eventHandler, event)
-                }
-            } else if (eventHandler.runMode == RunMode.INHERIT) {
-                runEventHandler(eventHandler, event)
-            }
-        }
-    }
+    abstract suspend fun dispatchEvent(event: Any)
 
     @JvmName("dispatchEvent")
     fun dispatchEventJava(event: Any) = runBlocking { dispatchEvent(event) }
 
-    fun dispatchEventAsync(event: Any): List<Deferred<Unit>> {
-        // Try not to switch context on non-handled events
-        // No need to check for `event` type as if it's in the map, then it's recognized
-        val handlers = map[event::class] ?: return emptyList()
-
-        val scope = coroutineScopesConfig.eventDispatcherScope
-        return handlers.map { eventHandler ->
-            scope.async { runEventHandler(eventHandler, event) }
-        }
-    }
-
-    private suspend fun runEventHandler(eventHandlerFunction: EventHandlerFunction, event: Any) {
-        try {
-            val (instance, function) = eventHandlerFunction.classPathFunction
-
-            /**
-             * See [CoroutineEventManager.handle]
-             */
-            val actualTimeout = eventHandlerFunction.timeout
-            if (actualTimeout.isPositive() && actualTimeout.isFinite()) {
-                // Timeout only works when the continuations implement a cancellation handler
-                val result = withTimeoutOrNull(actualTimeout.inWholeMilliseconds) {
-                    function.callSuspend(instance, event, *eventHandlerFunction.parameters)
-                }
-                if (result == null) {
-                    logger.debug { "Event of type ${event.javaClass.simpleName} timed out." }
-                }
-            } else {
-                function.callSuspend(instance, event, *eventHandlerFunction.parameters)
-            }
-        } catch (e: InvocationTargetException) {
-            if (event is InitializationEvent) {
-                //Entry point will catch exception as it is the one dispatching the initialization events
-                throw e.cause!!
-            }
-
-            if (e.cause is CancellationException) return
-
-            printException(event, eventHandlerFunction, e)
-        } catch (_: CancellationException) {
-            // Ignore
-        } catch (e: Throwable) {
-            printException(event, eventHandlerFunction, e)
-        }
-    }
-
-    private fun Collection<ClassPathFunction>.addAsEventListeners() = this
-        .requiredFilter(FunctionFilter.nonStatic())
-        .requiredFilter(FunctionFilter.firstArg(GenericEvent::class, BGenericEvent::class))
-        .forEach { classPathFunc ->
-            val function = classPathFunc.function
-            val annotation = function.findAnnotationRecursive<BEventListener>()
-                ?: throwInternal(function, "Function was asserted to have BEventListener but it was not found")
-
-            val parameters = function.nonInstanceParameters
-
-            val eventErasure = parameters.first().type.jvmErasure
-            if (!annotation.ignoreIntents && eventErasure.isSubclassOf<Event>()) {
-                @Suppress("UNCHECKED_CAST")
-                val requiredIntents = GatewayIntent.fromEvents(eventErasure.java as Class<out Event>)
-                val missingIntents = requiredIntents - jdaService.intents - config.ignoredIntents
-                if (missingIntents.isNotEmpty()) {
-                    return@forEach logger.debug { "Skipping event listener ${function.shortSignature} as it is missing intents: $missingIntents" }
-                }
-
-                // Cannot check for RawGatewayEvent as JDA is not present yet and there is no config for it
-            }
-
-            val eventParametersErasures = parameters.drop(1).map { it.type.jvmErasure }
-                // The main risk was with injected services, as they may not be available at that point,
-                // but they are pretty much limited to objects manually added by the framework, before the service loading occurs
-                .onEach {
-                    serviceContainer.canCreateService(it)?.let { serviceError ->
-                        throwArgument(
-                            classPathFunc.function,
-                            "Unable to register event listener due to an unavailable service: ${serviceError.toSimpleString()}"
-                        )
-                    }
-                }
-            @Suppress("DEPRECATION")
-            val eventHandlerFunction = EventHandlerFunction(classPathFunction = classPathFunc,
-                runMode = if (annotation.async) RunMode.ASYNC else annotation.mode,
-                timeout = getTimeout(annotation),
-                priority = annotation.priority,
-                parametersBlock = {
-                    //Getting services is delayed until execution, as to ensure late services can be used in listeners
-                    serviceContainer.getParameters(eventParametersErasures).toTypedArray()
-                })
-
-            classPathFunc.function.declaringClass.java.let { clazz ->
-                val instanceMap = listeners.computeIfAbsent(clazz) { hashMapOf() }
-
-                (eventTreeService.getSubclasses(eventErasure) + eventErasure).forEach {
-                    instanceMap.computeIfAbsent(it) { SortedList(EventHandlerFunction.priorityComparator) }.add(eventHandlerFunction)
-                }
-            }
-
-            (eventTreeService.getSubclasses(eventErasure) + eventErasure).forEach {
-                map.computeIfAbsent(it) { SortedList(EventHandlerFunction.priorityComparator) }.add(eventHandlerFunction)
-            }
-        }
-
-    private fun getTimeout(annotation: BEventListener): Duration {
-        if (annotation.timeout < 0) return Duration.INFINITE
-
-        return annotation.timeout.toDuration(annotation.timeoutUnit.toDurationUnit()).let {
-            when {
-                it.isPositive() && it.isFinite() -> it
-                else -> eventTimeout // Inherit from the (possibly user-provided) CoroutineEventManager
-            }
-        }
-    }
-
-    private fun printException(event: Any, eventHandlerFunction: EventHandlerFunction, e: Throwable) =
-        logger.error(e.unwrap()) {
-            "An exception occurred while dispatching a ${event.javaClass.simpleNestedName} for ${eventHandlerFunction.classPathFunction.function.shortSignature}"
-        }
+    abstract fun dispatchEventAsync(event: Any): List<Deferred<Unit>>
 }
