+++ src/main/kotlin/io/github/freya022/botcommands/internal/core/hooks/EventHooksProvider.kt
@@ -1,0 +1,46 @@
+package io.github.freya022.botcommands.internal.core.hooks
+
+import dev.minn.jda.ktx.events.CoroutineEventManager
+import io.github.freya022.botcommands.api.core.ICoroutineEventManagerSupplier
+import io.github.freya022.botcommands.api.core.JDAService
+import io.github.freya022.botcommands.api.core.config.BConfig
+import io.github.freya022.botcommands.api.core.config.BCoroutineScopesConfig
+import io.github.freya022.botcommands.api.core.service.ServiceContainer
+import io.github.freya022.botcommands.api.core.service.annotations.BService
+import io.github.freya022.botcommands.api.core.service.annotations.ServiceType
+import io.github.freya022.botcommands.internal.core.EventTreeService
+import io.github.freya022.botcommands.internal.core.service.FunctionAnnotationsMap
+import net.dv8tion.jda.api.hooks.IEventManager
+import org.springframework.context.annotation.Bean
+import org.springframework.stereotype.Component
+
+@BService
+@Component
+internal open class EventHooksProvider(
+    eventManagerSupplier: ICoroutineEventManagerSupplier,
+) {
+
+    private val originalCoroutineEventManager = eventManagerSupplier.get()
+
+    @Bean
+    @BService
+    @ServiceType(IEventManager::class)
+    internal open fun coroutineEventManager(
+        eventDispatcher: EventDispatcherImpl,
+    ): CoroutineEventManager {
+        return DispatcherAwareCoroutineEventManager(originalCoroutineEventManager, eventDispatcher)
+    }
+
+    @Bean
+    @BService
+    fun eventDispatcher(
+        config: BConfig,
+        coroutineScopesConfig: BCoroutineScopesConfig,
+        serviceContainer: ServiceContainer,
+        eventTreeService: EventTreeService,
+        jdaService: JDAService,
+        functionAnnotationsMap: FunctionAnnotationsMap,
+    ): EventDispatcherImpl {
+        return EventDispatcherImpl(config, coroutineScopesConfig, serviceContainer, originalCoroutineEventManager, eventTreeService, jdaService, functionAnnotationsMap)
+    }
+}
